Completed assignment

+++ Bill3.java
@@ -4,7 +4,7 @@ public class Bill3 {
 	    String name1 = args[0];
 		String name2 = args[1];
 		String name3 = args[2];
-		double sum = (args[3]);
+		int sum = (args[3]);
 		sum = Math.ceil(sum/3.0);
 		System.out.println("Dear " + name3 + ", " + name2 + ", " + name1 + ": pay " + sum + " Shekels each.");
 	}
